+++ benefits/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [typescript v2.0.1] benefits
 ### Releases
 - [NPM v2.0.1] https://www.npmjs.com/package/@wingspan/benefits/v/2.0.1 - benefits
+
+## 2023-12-11 01:26:27
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.125.2 (2.210.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.0.2] benefits
+### Releases
+- [NPM v2.0.2] https://www.npmjs.com/package/@wingspan/benefits/v/2.0.2 - benefits